Refactor WaveNet into NAM/wavenet/ (params, detail, model)
Move configuration types to params.h, implementation classes to detail.h,
and WaveNet/WaveNetConfig/JSON parsing to model.h and model.cpp. Slimmable and
tests include wavenet/model.h; CMake globs NAM/*/*.cpp for the new path.
Update Doxygen RST class paths. Format with clang-format.

Made-with: Cursor
