+++ tools/test/test_wavenet/test_layer1x1.cpp
@@ -7,7 +7,7 @@
 #include <stdexcept>
 #include <vector>
 
-#include "NAM/wavenet.h"
+#include "NAM/wavenet/model.h"
 
 namespace test_wavenet
 {
@@ -20,21 +20,21 @@ static nam::wavenet::_FiLMParams make_default_film_params()
 }
 
 // Helper function to create a Layer with default FiLM parameters
-static nam::wavenet::_Layer make_layer(const int condition_size, const int channels, const int bottleneck,
-                                       const int kernel_size, const int dilation,
-                                       const nam::activations::ActivationConfig& activation_config,
-                                       const nam::wavenet::GatingMode gating_mode, const int groups_input,
-                                       const int groups_input_mixin,
-                                       const nam::wavenet::Layer1x1Params& layer1x1_params,
-                                       const nam::wavenet::Head1x1Params& head1x1_params,
-                                       const nam::activations::ActivationConfig& secondary_activation_config)
+static nam::wavenet::detail::Layer make_layer(const int condition_size, const int channels, const int bottleneck,
+                                              const int kernel_size, const int dilation,
+                                              const nam::activations::ActivationConfig& activation_config,
+                                              const nam::wavenet::GatingMode gating_mode, const int groups_input,
+                                              const int groups_input_mixin,
+                                              const nam::wavenet::Layer1x1Params& layer1x1_params,
+                                              const nam::wavenet::Head1x1Params& head1x1_params,
+                                              const nam::activations::ActivationConfig& secondary_activation_config)
 {
   auto film_params = make_default_film_params();
   nam::wavenet::LayerParams layer_params(condition_size, channels, bottleneck, kernel_size, dilation, activation_config,
                                          gating_mode, groups_input, groups_input_mixin, layer1x1_params, head1x1_params,
                                          secondary_activation_config, film_params, film_params, film_params,
                                          film_params, film_params, film_params, film_params, film_params);
-  return nam::wavenet::_Layer(layer_params);
+  return nam::wavenet::detail::Layer(layer_params);
 }
 
 void test_layer1x1_active()
@@ -198,7 +198,7 @@ void test_layer1x1_inactive_bottleneck_mismatch()
   bool threw_exception = false;
   try
   {
-    auto layer = nam::wavenet::_Layer(layer_params);
+    auto layer = nam::wavenet::detail::Layer(layer_params);
   }
   catch (const std::invalid_argument& e)
   {
@@ -236,7 +236,7 @@ void test_layer1x1_post_film_active()
                                          nam::activations::ActivationConfig{}, film_params, film_params, film_params,
                                          film_params, film_params, film_params, layer1x1_post_film_params, film_params);
 
-  auto layer = nam::wavenet::_Layer(layer_params);
+  auto layer = nam::wavenet::detail::Layer(layer_params);
 
   // Set weights: conv, input_mixin, layer1x1, layer1x1_post_film
   // With bottleneck=channels=2:
@@ -314,7 +314,7 @@ void test_layer1x1_post_film_inactive_with_layer1x1_inactive()
   bool threw_exception = false;
   try
   {
-    auto layer = nam::wavenet::_Layer(layer_params);
+    auto layer = nam::wavenet::detail::Layer(layer_params);
   }
   catch (const std::invalid_argument& e)
   {
