+++ tools/test/test_wavenet/test_real_time_safe.cpp
@@ -9,7 +9,7 @@
 #include <string>
 #include <vector>
 
-#include "NAM/wavenet.h"
+#include "NAM/wavenet/model.h"
 #include "NAM/conv1d.h"
 #include "../allocation_tracking.h"
 
@@ -24,33 +24,31 @@ static nam::wavenet::_FiLMParams make_default_film_params()
 }
 
 // Helper function to create a Layer with default FiLM parameters
-static nam::wavenet::_Layer make_layer(const int condition_size, const int channels, const int bottleneck,
-                                       const int kernel_size, const int dilation,
-                                       const nam::activations::ActivationConfig& activation_config,
-                                       const nam::wavenet::GatingMode gating_mode, const int groups_input,
-                                       const int groups_input_mixin,
-                                       const nam::wavenet::Layer1x1Params& layer1x1_params,
-                                       const nam::wavenet::Head1x1Params& head1x1_params,
-                                       const nam::activations::ActivationConfig& secondary_activation_config)
+static nam::wavenet::detail::Layer make_layer(const int condition_size, const int channels, const int bottleneck,
+                                              const int kernel_size, const int dilation,
+                                              const nam::activations::ActivationConfig& activation_config,
+                                              const nam::wavenet::GatingMode gating_mode, const int groups_input,
+                                              const int groups_input_mixin,
+                                              const nam::wavenet::Layer1x1Params& layer1x1_params,
+                                              const nam::wavenet::Head1x1Params& head1x1_params,
+                                              const nam::activations::ActivationConfig& secondary_activation_config)
 {
   auto film_params = make_default_film_params();
   nam::wavenet::LayerParams layer_params(condition_size, channels, bottleneck, kernel_size, dilation, activation_config,
                                          gating_mode, groups_input, groups_input_mixin, layer1x1_params, head1x1_params,
                                          secondary_activation_config, film_params, film_params, film_params,
                                          film_params, film_params, film_params, film_params, film_params);
-  return nam::wavenet::_Layer(layer_params);
+  return nam::wavenet::detail::Layer(layer_params);
 }
 
 // Helper function to create a LayerArray with default FiLM parameters
-static nam::wavenet::_LayerArray make_layer_array(const int input_size, const int condition_size, const int head_size,
-                                                  const int channels, const int bottleneck, const int kernel_size,
-                                                  const std::vector<int>& dilations,
-                                                  const nam::activations::ActivationConfig& activation_config,
-                                                  const nam::wavenet::GatingMode gating_mode, const bool head_bias,
-                                                  const int groups_input, const int groups_input_mixin,
-                                                  const nam::wavenet::Layer1x1Params& layer1x1_params,
-                                                  const nam::wavenet::Head1x1Params& head1x1_params,
-                                                  const nam::activations::ActivationConfig& secondary_activation_config)
+static nam::wavenet::detail::LayerArray make_layer_array(
+  const int input_size, const int condition_size, const int head_size, const int channels, const int bottleneck,
+  const int kernel_size, const std::vector<int>& dilations, const nam::activations::ActivationConfig& activation_config,
+  const nam::wavenet::GatingMode gating_mode, const bool head_bias, const int groups_input,
+  const int groups_input_mixin, const nam::wavenet::Layer1x1Params& layer1x1_params,
+  const nam::wavenet::Head1x1Params& head1x1_params,
+  const nam::activations::ActivationConfig& secondary_activation_config)
 {
   auto film_params = make_default_film_params();
   // Duplicate activation_config, gating_mode, and secondary_activation_config for each layer (based on dilations size)
@@ -65,7 +63,7 @@ static nam::wavenet::_LayerArray make_layer_array(const int input_size, const in
     std::move(activation_configs), std::move(gating_modes), head_bias, groups_input, groups_input_mixin,
     layer1x1_params, head1x1_params, std::move(secondary_activation_configs), film_params, film_params, film_params,
     film_params, film_params, film_params, film_params, film_params);
-  return nam::wavenet::_LayerArray(params);
+  return nam::wavenet::detail::LayerArray(params);
 }
 
 // Helper function to create LayerArrayParams with default FiLM parameters
@@ -91,7 +89,7 @@ static nam::wavenet::LayerArrayParams make_layer_array_params(
 }
 
 // Helper function to create a Layer with all FiLMs active
-static nam::wavenet::_Layer make_layer_all_films(
+static nam::wavenet::detail::Layer make_layer_all_films(
   const int condition_size, const int channels, const int bottleneck, const int kernel_size, const int dilation,
   const nam::activations::ActivationConfig& activation_config, const nam::wavenet::GatingMode gating_mode,
   const int groups_input, const int groups_input_mixin, const nam::wavenet::Layer1x1Params& layer1x1_params,
@@ -106,7 +104,7 @@ static nam::wavenet::_Layer make_layer_all_films(
                                          gating_mode, groups_input, groups_input_mixin, layer1x1_params, head1x1_params,
                                          secondary_activation_config, film_params, film_params, film_params,
                                          film_params, film_params, film_params, film_params, head1x1_post_film_params);
-  return nam::wavenet::_Layer(layer_params);
+  return nam::wavenet::detail::Layer(layer_params);
 }
 
 // Test that pre-allocated Eigen operations with noalias() don't allocate
@@ -739,7 +737,7 @@ void test_layer_post_activation_film_gated_realtime_safe()
     inactive_film, // _1x1_post_film
     inactive_film // head1x1_post_film
   );
-  auto layer = nam::wavenet::_Layer(layer_params);
+  auto layer = nam::wavenet::detail::Layer(layer_params);
 
   // Set weights - Order: conv, input_mixin, 1x1, then FiLMs
   // NOTE: In GATED mode, conv and input_mixin output 2*bottleneck channels!
@@ -846,7 +844,7 @@ void test_layer_post_activation_film_blended_realtime_safe()
     inactive_film, // _1x1_post_film
     inactive_film // head1x1_post_film
   );
-  auto layer = nam::wavenet::_Layer(layer_params);
+  auto layer = nam::wavenet::detail::Layer(layer_params);
 
   // Set weights - Order: conv, input_mixin, 1x1, then FiLMs
   // NOTE: In BLENDED mode, conv and input_mixin output 2*bottleneck channels!
@@ -1196,7 +1194,7 @@ void test_process_3in_2out_realtime_safe()
   }
 }
 
-// WaveNet::process() with optional post-stack head (multi-layer PostStackHead) must not allocate or free.
+// WaveNet::process() with optional post-stack head (multi-layer detail::Head) must not allocate or free.
 void test_process_with_post_stack_head_realtime_safe()
 {
   const int input_size = 1;
@@ -1223,7 +1221,7 @@ void test_process_with_post_stack_head_realtime_safe()
                             std::move(dilations), activation, gating_mode, head_bias, groups, groups_input_mixin,
                             layer1x1_params, head1x1_params, nam::activations::ActivationConfig{}));
 
-  nam::wavenet::WaveNetHeadParams head_params;
+  nam::wavenet::HeadParams head_params;
   head_params.in_channels = 1;
   head_params.channels = 1;
   head_params.out_channels = 1;
@@ -1243,7 +1241,7 @@ void test_process_with_post_stack_head_realtime_safe()
   std::unique_ptr<nam::wavenet::WaveNet> condition_dsp = nullptr;
   auto wavenet =
     std::make_unique<nam::wavenet::WaveNet>(input_size, layer_array_params, head_scale, with_head,
-                                            std::optional<nam::wavenet::WaveNetHeadParams>(std::move(head_params)),
+                                            std::optional<nam::wavenet::HeadParams>(std::move(head_params)),
                                             std::move(weights), std::move(condition_dsp), 48000.0);
 
   const int maxBufferSize = 256;
